Fix for bug 612002
Undos previous fix that changed the compiler to always emitted a generated property name of __result for EndInvoke.

The new change only adds the underscores if there is a name conflict. In this way, we stay metadata compatible with the prior native compiler.

Also added same fix to generated parameters for BeginInvoke, callback and object.

Added test cases. (changeset 1393061)