Add an analyzer to check for changes to the public API.
This change adds an analyzer that raises errors when modifying the public APIs.

The idea here is that each project contains a file, PublicAPI.txt, that lists the full names of the types and members that make up the public API of that assembly. When the diagnostic encounters a public symbol, it checks if the name is in the list. If not, it creates a diagnostic. While this does not prevent changes to the public API, it will make them very obvious in code reviews and source control history as the only way to get rid of the error is to update PublicAPI.txt.

At the end of compilation, we also check for symbols that were declared in PublicAPI.txt but were not found in the project, or were not public, and report errors about those as well.

The bulk of the change is actually in the fixer, as I choose to implement a custom FixAllProvider in order to test out that feature. (changeset 1396119)