DevDiv #1105950: AQNs in CDI
Assembly-qualified names in custom debug info should be produced using the same code path as assembly-qualified names in emitted code.  This fixes an unhandled case (pointers as type arguments).

This also brings us closer to the native behavior: we had started emitting the assembly name for types in the current assembly, which the native compiler did not.

CR: alekseyt; tomat (changeset 1397706)