[Port changeset 1110045 to RC branch]
Fix for bug 1110045: Disable concurrent execution of analyzer actions

Compiler AnalyzerDriver processes compilation events sequentially, but executes analyzer actions for each event in parallel. This doesn't work well with analyzers that are not thread-safe, and we don't expect everyone to write thread-safe high performant analyzers. Until, we add some support to analyzers to allow it to specify whether or not it is a thread-safe analyzer, we should disable concurrent execution of analyzer actions for every analyzer.
This change primarily does couple of things:
1) Group analyzer actions by analyzer in the AnalyzerDriver.
2) For symbol declared events, create a task chain per analyzer with continuations added for executing actions for that analyzer. For other compilation event kinds, we can just execute actions per-analyzer in a single task.
This change is *expected* to increase the overall build time as we lose concurrency of analyzer execution, but hopefully that will be pulled back when we add support for registering an analyzer as thread-safe analyzer and mark all our analyzers as thread-safe. (changeset 1400889)