Port changeset 1401689 from Main to Dev14RC
MEF sometimes ends up populating returned arrays for IOrderableMetadata.After and IOrderableMerada.Before with null entries. This used to be harmless until recently - but is no longer harmless after the recent API change that made the ExportCodeFixProviderAttribute.Name property for code fix providers optional.

If a code fix provider does not specify a Name via this attribute, then the above MEF behavior means that we end up seeing cycles in our extension ordering implementation. The previously harmless nulls inside ExtensionOrderAttribute.Before and .After are perceived to be the same thing as the code fix providers with null names - and because MEF can return null entries in both the .Before and .After collections, extension orderer thinks that we have a cycle with the same (null-named) fix provider being ordered both before and after some other fix provider. The orderer then throws an exception and we refuse to offer any fixes to the user in the lightbulb. Worse still, the only thing that the user can do to recover from this situation is to figure out and remove the fix provider that has a null Name from the mix.

There are two issues that needed fixing here -

1. Fix providers with null names should never cause accidental cycles like above. We filter nulls out of collections returned from MEF to avoid such cycles.

2. Even if someone intentionally created a cycle (by placing same extension before and after another via [ExtensionOder()] attributes), we should never crash. We should break the cycle but avoid the crash - light bulb should continue to offer all available fixes even if a conflict is detected in the ordering.

Also discovered that the CSharpConvertToAsyncMethodCodeFixProvider was using incorrect name + that EncapsulateField refactoring was not listed in PredefinedCodeRefactoringProviderNames. Fixing both issues. (changeset 1401717)