Compat: Suppressing branch-coalescing optimizations in certain rare cases where it can result in extra work in JIT or even expose JIT bugs.
== Consider the following code:

        try
        {
            try
            {
                // SOME CODE
            }
            catch (Exception ex)
            {
                Console.WriteLine("catch1");
            }
        }
        catch (Exception ex)
        {
            Console.WriteLine("catch2");
        }

=== It's IL equivalent is

        try
        {
            try
            {
                // SOME CODE
                leave LABEL1;
            }
            catch (Exception ex)
            {
                Console.WriteLine("catch1");
                leave LABEL1;
            }
LABEL1:
                leave LABEL2;
        }
        catch (Exception ex)
        {
            Console.WriteLine("catch2");
                leave LABEL2;
        }
LABEL2:

Note that LABEL1 is technically redundant. All branches to LABEL1 can as well point directly to LABEL2.
Indeed, branch optimizer notices this fact and morphs the code into more compact:

        try
        {
            try
            {
                // SOME CODE
                                leave LABEL2;
            }
            catch (Exception ex)
            {
                Console.WriteLine("catch1");
                                leave LABEL2;
            }
        }
        catch (Exception ex)
        {
            Console.WriteLine("catch2");
                    leave LABEL2;
        }
LABEL2:

While semantically it is a valid transformation, it creates two issues related to handling of Thread.Abort :

1) if Thread.Abort happens while running SOME CODE, then the first catch will catch it, however since it is not explicitly suppressing the propagating of the thread abort, the exception must be rethrown by the runtime. The inconvenient part here is that the exception must be rethrown when running some code outside of the inner try, but inside the outer try. Since there is no such code in this IL, JIT must detect this situation and do some code injection that is similar to basically undoing the optimization that C# compiler did.

Considering that the only thing that we gain here is removing one LEAVE instruction at the cost of requiring JIT to essentially put it back, it does not seem that the optimization is very profitable.

2) The part where JIT detects such cases and special cases them is _very_ delicate. There appear to be a variety of bugs in different JITs which result in missing this case in one form or another resulting in some handlers not being run as expected.

Indeed, it has been always possible and is still possible to introduce this case by hand via explicit goto statements and cause bad behavior. These issues would be not new and only JIT developers can fix them.

However introducing this case as a result of optimizations causes compatibility issues where some code that formerly worked no longer works when recompiled with Roslyn, because there are cases that Dev12 compilers did not optimize.

To mitigate the compat issues we will explicitly suppress label-coalescing optimizations when they may result in a label to "move out" of a region protected by a catch.

 (changeset 1402671)