Show the correct effective severity for diagnostics in Solution Explorer.
It turns out the code in Solution Explorer to calculate the effective severity for a diagnostic (and thus show the correct icon) was using the diagnostic's enabled-by-default flag and the compilation's general diagnostic option in the wrong way.

If a diagnostic is disabled by default, then the only thing that can turn it on is a specific diagnostic option, not the general diagnostic option. Also, the general diagnostic option can only turn a warning into an error, or suppress a warning or info. (changeset 1403379)