Change the Color Color rule in VB for backwards compatibility with Dev12.
Before this change, Roslyn used a slightly different Color Color rule for Visual Basic than the Dev12 compiler. Where the Dev12 compiler's rules would never result in an unresolvable ambiguity, the Roslyn compiler's could, particularly if the LHS required late binding or had errors. The Color Color rules have been therefore been adjusted to make decisions in the same manner as Dev12 for late-bound cases (where we must be compatible). The new rules are given below.

If the base expression of a member access expression is a simple name and binds to a constant, field, property, local variable or parameter whose type has the same name, then the base expression can refer either to the member or the type. This can never result in ambiguity because the members that can be accessed off of either one are the same.

In the case that such a base expression binds to an instance member but the binding occurs within a context in which "Me" is not accessible, the expression instead binds to the type (if applicable).

If the base expression cannot be successfully disambiguated by the context in which it occurs, it binds to the member. This can occur in particular in late-bound calls or error conditions.
***NO_CI***
 (changeset 1403686)