[Non-Shipping] Merge nuget package authoring to RC branch as we want to release nuget packages for Roslyn analyzers when RC ships.
Creating one nuget package for FxCopAnalyzers that contains both C# and VB and an update install script that knows to install the right analyzer for the target right target project.

Creating a similar package for CodeAnalysis analyzers. Adding a dependency to this package from the CodeAnalysis.Common package so that any time a codeanalysis nuget is installed, the analyzers are installed as well.
 (changeset 1404502)