The Open directory contains several projects that rely on having a compatible version of Visual Studio 2015 installed in order to build properly.   At this time there is no version of Visual Studio that will work over the long term that we can ask customers, and Jenkins, to move to in order to edit and build Roslyn.sln on master.  Our dependency on the VS SDK just evolves too rapidly.
After some discussion I decided to put two solution files into the repo:

     - Roslyn2013.sln:  This solution is usable on a machine with only Visual Studio 2013 installed.  This unfortunately won't contain all of the new items we are releasing but at least gives customers a stable base to iterate on.
     - Roslyn.sln: This solution contains every project we've released and will work with some builds of 2015.  Likely it will work with the most recent CTP but I have not verified that (just don't have the right setup to do so).

The long term plan here is to always be able to edit, build and debug the master branch with an RTM version of Visual Studio.  But until we hit RTM (or really RC) of Visual Studio 2015 we are going to have a hybrid solution to keep both us and customers productive.   (changeset 1405691)