[Porting to RC] Make AsyncQueue and CompilationEvents internal.
(checking in on behalf of Jared)

After some discussion between Neal, Sri and myself we decided to make AsyncQueue and CompilationEvent internal. The justification for this was: - At the time there is no need for this type to be public. Ideally the IDE and Compiler would be structured in a way that making it public was needed but for now it is not done that way. - AsyncQueue feels like a type which should be in BCL, not Roslyn. - No concrete need for making this public at the moment. Given that we decided to make it internal for now. One of the main hopes here is we can convince the BCL to take ownership of the AsyncQueue. If we make it public now, even if they add such a type we'd never get to consume it.
 (changeset 1406171)