Allow async taggers to specify what sort of tracking span behavior they want.
This affects what happens to tags for previous text snapshot versions that are mapped forward (in the interim period before we get the new up to date tags).

Most features just have EdgeExclusive behavior.  However, some features, like TypeScript classification, want EdgeInclusive behavior so that if you type at the edge of an existing tag, the next text gets the old classification and does not flash.
***NO_CI***
 (changeset 1406469)