Ignore assembly names when resolving symbols in Peek.
This is because peek uses symbolkeys to resolve symbols between compilations which could potentially be from different frameworks but symbolkeys do not contain enough information to deal with typeforwarding situations. E.g: a portable project and a non-portable project may both resolve the same Int32 symbol to System.Runtime and mscorlib.

Other features such as Metadata as source, Find References already do this. This change updates Peek to do the same thing. (changeset 1407182)