Add in the missing files
Our script for replaying TFS changes on our Git repo missed a few files in the last change.  TFS labeled these files as "undeleted" which the script doesn't appear to handle correctly.  This isn't surprising as this comes about when TFS applies a roll back of a roll back (not very common).

Manually adding these files to the Git repo to unblock the build.