Remove some unnecessary ProjectReferences
We have some project references between projects that are either completely unnecessary or almost so, which have the nice side effect of creating build warnings due to version mismatches. Remove most of them, and move a single unit test helper next to the unit tests that are using it. (changeset 1408938)