Add an analyzer that can detect unused declarations.
this is compilation end analyzer. this is supposed to squiggle name and fade out whole code block when it detects unused declarations. but for now, it only reports the issue in error list and not in editor due to engine limitation.

we will use this analyzer to drive making compilation end analyzer experience better. so that we can get things like squiggle, fading out or code fix work with compilation end analyzer. (changeset 1409499)