Fix Bug 1087167 - Format AttributeLists of a parameter appropriately
AttributeLists of the Parameter should be left untouched by the VB PrettyLister. The AttributeLists of the other statement is formatted in such way that the first attributeList of the lot is formatted with expected indentation and the rest of the attributeList anchors to the first attributeList and moves accordingly. (changeset 1409606)