Fix empty array tests in Jenkins
While our internal runtime contains the new Array.Empty member, the runtime used by tests in Jenkins does not.  This caused the Jenkins tests to begin to fail once we updated our empty array suites to always expect the member to be present.  I changed the suites to test for the member and change their behavior based in its presence or absence.

 (changeset 1409795)