Show line information for diagnostics in PublicAPI.txt.
If a PublicAPI.txt file includes a symbol that is not found in the assembly we will create an error stating that an API has been deleted. We put this error in the PublicAPI.txt file, but do not currently include the line number.

With this change we now search for the proper line, and include it in the diagnostic. (changeset 1411808)