Provide advanced matching for dotted names in NavigateTo.
No, if you search for something like CS.SK we'll find Microsoft.CodeAnalysis.CSharp.SyntaxKind

We do this by breaking up the original pattern by dots (so, in this case, we'd have "CS" and "SK").  We then do our normal search, looking for symbols that match the last name piece (in this case "SK").  When we find one, we then walk up it's container seeing if each parent container name matches each respective preceding pattern component.  If all pattern parts match, then the symbol is a match and it is returned.
***NO_CI***
 (changeset 1411849)