Add a small cache of AbstractSyntaxContext
AbstractSymbolCompletionProvider computes a SyntaxContext and passes
that around while computing CompletionItems. Many CompletionProviders
derive AbstractSymbolCompletionProvider in order to share the logic for
linked files completion. This means that we'll repeatedly compute the
same SyntaxContext for a given Document and position. This shows up in
the 2-core typing tests. To avoid this, we'll simply cache the most
recent Document/position/SyntaxContext and use it if it matches the
current request.