Improve project load/unload performance
When a project with N files is loaded or unloaded, the project will be
forked N times, each time adding or removing one of the N files. For
each of those forks, we need to update the Solution's map of linked
files. The process of updating the linked files map was previously
O(N^2), resulting in an overall O(N^3) project load/unload process. This
change makes makes the linked file map calculation O(1) and therefore
the project load/unload process O(N).

Updating the linked files map was previously O(N^2) because it tried to
account for arbitrary project edits that might have added any number of
files and removed any number of files. However, all project-related
forking actually happens at either the level of the entire project
(which was already special-cased) or at the level of an individual file.
We now handle the case of individual file adds/removes efficiently.

This fixes internal issue 1112399