Support caching multiple documents
Keep a Dictionary of Document -> Task<SyntaxContext> and a position.
Every time we start computing items, check if the Documents to compute
against are all keys in this dictionary. If they are, return cached
tasks. Otherwise, clear the dictionary and immediately add the new set
of Documents as keys. As we iterate over the set of Documents to compute
completion for, we update the cache to point to the Task returned by
CreateContext.