Fix RunTests
It is apparently possible for a test to exit with an error code, produce
no HTML file and no output.  In that case the call to Aggregate was
failing with an exception.  Change it to use ToArray +
File.WriteAllLines.

This won't fix the underlying problem but will at least identify the
test that is failing.