Exclude mscorlib from Emit test conflict checking
In the emit tests we currently check every module to see if we're
emitting an assembly with a name we've already seen. PeVerify
requires that we create a new AppDomain for every conflict we find.
However, mscorlib will always produce a conflict, so this ignores
mscorlib when checking for a conflict and doesn't create a new
AppDomain if it is found during emit.