Enable 64 bit test runs
This change allows provides the option of running our tests in 64 bit
mode.  Simply invoke BuildAndTest.proj in the following manner:

> msbuild /v:m /m BuildAndTest.proj /p:Test64=true

There are a set of tests today that don't run correctly in 64.  All have
been annotated with the trait

> Require32=true

We need to push on these to get a clean 64 bit run.  But short term this
will at least give us protection against regressions.