Fix Issue #146
It turns out that there was a fundamental problem with how parameter
names were handled in VB Code Model. Essentially, the CodeParameter
object was created with the modified identifier of the name. So, it
included the ? or () if they were declared and the type character if it
was declared. This made a bit of a mess when it came time to look of the
underlying node. This should fix several problems (not all reported), so
I went ahead and added several unit tests.