Reduce the amount of work done in the pattern matcher.
Previously, the way one used a patter matcher was the following:

    var matcher = new PatternMatcher();
    Loop {
        var result = matcher.MatchPattern(candidate, pattern)
    }

Because of this, each time you called MatchPattern, the matcher would have to deconstruct the
pattern into its relevant components (for example, splitting it on dots and decomposing words out
of it).  Now, this decomposition was cached.  However, that still meant checking the cache on each
call to MatchPattern.  This check involved a lock as well as a dictionary lookup.

The new way to use a pattern matcher is:

    var matcher = new PatternMatcher(pattern);
    Loop {
        var result = matcher.MatchPattern(candidate)
    }

This approach means that we only need to decompose the pattern once.  From then on we can reuse
that decomposition for each value we test against.