Support setting the active rule set from Solution Explorer
Fixes #90.

Currently you have to open a project's Code Analysis property page to
pick the rule set file in effect for that project. With this change we
allow the user to simply right-click on a .ruleset file in the project
in Solution Explorer and choose "Set as Active Rule Set".