add property bag to diagnostic and remember origin of the diagnostic
this change contains 2 changes. one is adding property bag to diagnostic and the other is putting origin of the diagnostic in the property bag.

the property bag is added based on users feedback. this basically lets users add a diagnostic specific information.

second change is for error list. we would like to order errors better based on the origin of the errors (syntax, declaration, others and etc). two changes are together because the second one takes advantage of the property bag we just added.

...

this is a port of CR I had in TFS to github. checkin is delayed due to sign off - this contains public API change.